Added ability to select which characters to practice.
This lets you remove ᖤ, ᖢ, and ᖠ if you're practicing a dialect that
doesn't use them, or to drill specific sets you're having trouble with.
I might add the Nunavik "ai-pai-tai" column in the future so it would be
useful for that too.

I want to clean up the UI for this to a dual listbox system, but this is
a functional start.